Improvements to the Responses component. Fix issues with maps used to generate the reponse row items.
